:bug: Fix ownership of polkitd rules dir
Due to a bug on earthly it seems that some of the directories for
polkitd are not being copied under the proper user. This results into
polkit not being able to read the dir properly and the rules that come
with it.

This patch works around it by installing an oem file that wil restore
the proper ownership of those dirs

Signed-off-by: Itxaka <itxaka@spectrocloud.com>