:robot: Adapt tests to immucore
Before immucore 0.0.15 we were ignoring this errors and the dag
continued with no issues.

Now, the dag requires the previous step to pass, so this is failing.

This reworks the test to make sure we mount /var first on the test as we
cannot try to mount something under /var if we havent mounted /var first

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>