:seedling: Generic arm64 isos (#1291)
:seedling: Generate arm64 generic isos

 - Use latest osbuilder tools image as it contains fixes for arm64 isos
 - Add a small job to build an iso as part of the image arm workflow
 - Rework /boot/vmlinuz linking so it works on x86 and arm64

This also adds an iso-remote target for earthfile that allows to build
an iso from a remote rootfs in the shape of a docker artifact. This
allows for easy reproduce of isos based on existing artifacts instead of
having to rebuild the whole thing. You point to the artifact and it
spits an iso really fast. Also works with arm64 artifacts so we can
easily build a local arm64 iso for testing for pushed artifacts.

This PR makes it so generation of isos for arm64 works.

Signed-off-by: Itxaka <itxakaserrano@gmail.com>