:bug: Fix persistent machine-id
Looks like we were doing it on the same stage which made no sense as by
that time there may not be a machine-id yet.

Now the machine id is splitted in two. If we have a stored machine ID,
during initramfs we copy it over /etc so systemd doesnt recreate one and
during fs if we dont have a stored machien id, we save the one
generated.

This results in the machine-id being persistent across reboots and as a
side effect things like getting the same dhcp ip on reboot now works out
of the box.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>