:bug: Merge grub yip config
With yip 1.x things can run at the same level as its now using a DAG.
Problem is with the grub stuff it was 2 files trying to mount and
unmount the state dir. That worked as expected before as it run one
after the other but now they can run in parallel as they dont depend on
each other.

This PR merges both configs into one to avoid mount/umount several times
and because both deal with grub stuff.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>