:bug: Mount bpf system once we boot
If we do it initramfs we are trooubling the chroot mounts that the stage
uses to run the config files. Nothing should be mounted in the binded
paths during that stage to avoid issues as after the stage is finished,
the chroot is removed fully, so those mounts would get lost.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>