:robot: Notify of main failures via slack
This introduces a way for the main branch to notify in the slack channel
when failures happen on the main branch. Currently we have no visibility
over what happens on the main branhc unless we check manually. And as we
are moving more jobs to be run only in the main branch we need to keep
up to date on those failures.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>