:robot: Fix+move image size test (#1648)
Moves the image size check into the acceptance tests.
Fixes to it takes into account the new image sizes. Before we were
overriding the image sizes with the elemental config yaml and setting
them to 3000 but now if not set via cloud config they will get set to
the default in kairos-agent which is 3072 and its a constant for all
default images.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>