:art: Drop files-iso folder from iso building
Package should provide the sane grub.cfg file directly and we should not
need to override it.
This bumps the repos to get the latest efi artifacts and grub packages
that provide a sane grub.cfg for both efi and bios, which results in
dropping the iso overlay that we had to add to make the isos work.

This is now provided by the package directly like it provided the efi
config

Signed-off-by: Itxaka <itxaka@kairos.io>