Remove failing check
Permissions seem to have changed to 770 which is what an installed system has anyway.

The 700 permissions seems to originate here:
https://github.com/kairos-io/kairos/blob/948f4b72274f6f561eacd9a90c6a36f53bcb86de/overlay/files/system/oem/10_accounting.yaml#L50

although I can't tell if that script was run when building the uki artifacts.

In any case, since we don't even know what the right permissions should
be, we don't need to check this.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>