Bundle luet versions in the framework
Use the new luet 0.35 which provides a way of storing the installed
artifacts in the system to bundle the versions file with the framework
files so we always have a reference for what was installed in the
framework and what versions.

Also reworks the CI pipeline for the bump repos to provide a diff of
changes to **installed** packages, not to the full repo unfortunately,
as part of the PR creation so its easily visible what changed that
directly affects kairos

By cleaning at the end of the actual targets we carry the luet database
to the end of the target so we can generate ALL the packages for
whaverer artifact we are building so we dont need to mangle the
different versions.

Either the framework-image or the core/standard base-image will provide
a final /etc/kairos/versions.yaml which provider all the artifacts
installed by luet

Signed-off-by: Itxaka <itxaka@kairos.io>