Fix test that wasn't testing anything
because it was returning success too early since the checks were too
loose or even wrong.

E.g. we didn't even wait for the Pods to be ready (wrong test condition)
and we only tested that there is a "v" in the version which was true
even before the upgrade.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>