:art: Bundle bootargs with framework images
Currently bootargs is not provided with the framework images so its
missing a very important piece to boot which results into BYOI being
broken.

This patch adds the bootargs.cfg to the framework image. Also the
config.txt to the rpi images.

Also drops the kernel/initramfs vars from the bootargs as they are set
to the default now in the package.

Also moves the luet clean under the package list generation as it
requires the database of luet to list the packages.

Signed-off-by: Itxaka <itxaka@kairos.io>