:art: Use a single framework image
With changes to the packages we now generate the proper service files on
the fly so we dont need to have those package separated.

We now have 2 framework images, a generic one used for ALL flavors and a
fips one, only provided for external consumers as we dont publish any
fips enabled artifacts.

We also bring both the systemd and openrc package together as they dont
affect each other anymore so we can have a single framework.

This patch also removes the service enablement as its not needed
anymore, it done via config files

Reworks the framework jobs to just build and publish the 2 different
framework image (generic and fips)

Signed-off-by: Itxaka <itxaka@kairos.io>