:robot: Fix some tests failing for k8s and run install under qemu
Fix tests missing the k8s version due to it being bumped on the
packages.
It will now try to use the usual method and if nothing is found resort
to a simple ls to get the latest version downloaded of the iso.

Alos make the install test run under qemu instead of vbox which is
terrible and slow and fails half of the time.

Signed-off-by: Itxaka <itxaka@kairos.io>