Revert to the old test logic
which, in case of errors, will fail with a timeout but kairos service
doesn't allow us to know when the installer is done. That's because
the service has `RemainAfterExit=yes` and even if that was removed, the
kairos process waits for user input to finish:

```
Installation completed, press enter to go back to the shell. [Yes/no/all/cancel]
```

so there is not sure way to tell whether the kairos installer failed
with or without errors.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>