Catch agent issue 174 (#2017)
* Check for failures faster and make sure bundles don't break

It should fail now because of this fix needed:

https://github.com/kairos-io/kairos-agent/pull/174

After bumping to the fixed kairos-agent version, the test should be
green again.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* :arrow_up: Update repositories

Signed-off-by: GitHub <noreply@github.com>

* Bump osbuilder to produce a livecd with tty1 as default console

Grub configuration for livecd (iso) is installed here:
https://github.com/kairos-io/osbuilder/blob/89b014b2b2a1f27bd0c0c5269cecec2a1b8f37bf/tools-image/Dockerfile#L25

and it repositories were bumped in osbuilder to use the latest and
updates grub configuration where `console=tty1` is set **after**
`console=ttyS0` thus making it the default console for systemd services.

Also:

- remove not used bootargs.cfg file (has been moved to the "packages"
  repo)
- Change to the DATASOURCE check which is what we need in reality (not
  CLOUD_INIT which was indirectly checking that DATASOURCE would be set
  by Earthly)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Double the waiting time until installation is finished

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Triple the timeout (it passses locally, just needs time)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Bump luet repository to get kairos-agent 2.4.2

which has this fix: https://github.com/kairos-io/kairos-agent/pull/174

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Revert to the old test logic

which, in case of errors, will fail with a timeout but kairos service
doesn't allow us to know when the installer is done. That's because
the service has `RemainAfterExit=yes` and even if that was removed, the
kairos process waits for user input to finish:

```
Installation completed, press enter to go back to the shell. [Yes/no/all/cancel]
```

so there is not sure way to tell whether the kairos installer failed
with or without errors.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

---------

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>
Signed-off-by: GitHub <noreply@github.com>
Co-authored-by: jimmykarily <jimmykarily@users.noreply.github.com>