:sparkles: Enable secureboot for all flavors (minus alpine)
This PR brings new deps, bumps builders and agent to bring secureboot in
all flavors but alpine (more ont hat below)

Currently to enable secureboot in a given X flavor we need to make sure
to use the artifacts from that falvor specifically as they are all
signed with the distro owner key, so the chain of validation works as
expected.

shim (dual signed Microsoft and distro owner) -> grub (distro owner
signed) -> kernel (distro owner signed)

So unless there comes a time where the shim is signed by Microsoft and
ALL the distro owners (so you can chainload all grubs and kernel signed
by all of them) we need to specifically use the distro artifacts so the
signatures match.

This PR brings a new enki version(inside osbuilder) which builds the iso
using the specifci distro artifacts (instead of generic ones as we used
to do).

It also makes sure that all distros include the grub and shim signed
artifacts in the base images.

Also brings a new framework that drops the generic grub artifacts (not
used) and it also brings a new agent that uses those new artifacts
during installation instead of the generic ones.

Unfortunately Alpine does not provide signed artifacts so SecureBoot is
not supported under Alpine until they provide those. We still use our
generic artifacts to enable uefi support for Alpine

Signed-off-by: Itxaka <itxaka@kairos.io>