Remove double brackets from Earthly (#2541)
Under the hood, Earthly calls `sh` to execute control flow. Unfortuantely this is prone to errors because the underlying container can either use a posix compliant `sh`, or a different one e.g. `bash`. One example of the former is Ubuntu:

```
/bin/sh: 1: [[: not found
```

To make matters worse, Earthly doesn't halt because of this failure, so the process finishes successful but build something unexpected.

fixes #1759

Signed-off-by: Mauro Morales <mauro.morales@spectrocloud.com>