:bug: Install cryptsetup for all arches in opensuse
Seems like we were only installing cryptsetup in amd64 and immucore
dracut install requires cryptsetup. This caused the list of required
binaries in the initramfs to fail in the middle and not install the next
ones.

So in the case of arm64 (rpi) this mean that binaries were installed up
to the cryptsetup, then it failed and didnt continued installing the
required binaries, which were the ones being used to expand the
partition to max size on boot.

This makes cryptsetup part of the common install for both arches

Signed-off-by: Itxaka <itxaka@kairos.io>