:penguin: Disable pcrlock for all systemd distros
the pcrlock service will generate a policy on boot which when adding our
measurements policy, will also be read and clash. You can only either
use a pcrlock policy or a measurements policy.

This disables the service until the pcrlock stuff is more stable and we
can move to use it directly

Signed-off-by: Itxaka <itxaka@kairos.io>