:bug: Empty machine-id instead of removing it
We used to rely on a hidden feature of sysinfo that created a filled
machine-id during initramfs so systemd did not trigger the first boot
service.

Now on the newer agent versions, that hidden functionality has gone away
so if we remove it teh file, systemd will think that this is a new
system and will do some work behind our back initializing and break
things around.

So we need to create en empty machine-id file so systemd knows that we
dont want to run the first boot services/target

Signed-off-by: Itxaka <itxaka@kairos.io>