:wrench: Allow testing overlya files branches (#2791)
Instead of waiting for them to trickled down, we can easily test them
and integrate them intot he rootfs.

This provides a OVERLAY_FILES_DEV_BRANCH argument that can be set to a
branch or commit of the packages repo in order to test ongoing branches
or single commits directly.

Currently does not remove the existing files so anything that removes
files cannot be tested by this as we cant now in advance which files to
delete, so it only works by overwriting the files or adding new ones.

Signed-off-by: Itxaka <itxaka@kairos.io>