Fix test printing old value for debugging
ShouldNot(Equal(currentVersion), out) is a method call and as such, it
received the value of `out` at the time of the call. That was not the
updated value after the `Eventually` timeout had passed but rather the
value from the previous call that populated it.
This made the test output very confusing because, even though the
upgrade Pod was erroring out, the output shows the cluster being only
some seconds old and the pod still being created (ouch!)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>