:bug: Do not bindly install all tpm2 tools
We were going a bit wide and installing everything under tpm2-*
This brougth a userspace daemon for tpm2 which was failing to start in
non-tpm enabled nodes, thus failing the systemd status and blocking boot

We added a workaround to time out the service but we should directly not
have that daemon in there. All our tpm2 stuff uses the in kernel tpm
calls as it needs to be really early so the daemon is useless in our
workflows.

Signed-off-by: Itxaka <itxaka@kairos.io>