:robot: Increase the UKI VM resources
It looks like during upgrade the resources for the VM can be exhausted
which kills the machine in the middle.

Now that we are running the UKI tests in our own workers we can increase
a bit the minimun resources for the test VMs.

This patch increases the memory to 6Gb from 4gb and the CPU from 2 to 4

Signed-off-by: Itxaka <itxaka@kairos.io>