Use kairos-init (#3130)
This uses Kairos init to build all images. X86, arm and UKI images are all build with init with a single docker file. Artifacts are then build from those, like isos or raw images

Also drops a bit of earthly ik the build process to directly build with docker and call aurora directly as well, so we have 1 less layer of complexity. Ideally with this, only docker is needed locally to generate final artifacts, upgrade images, raw images, UKI artifacts (EFI, upgrade and iso) and so on. Init + aurora substitutes the earthly pipeline for building.

Testing still relies on earthly for the ginkgo tests.


---------

Signed-off-by: Itxaka <itxaka@kairos.io>