Add grype ignore file
grype is not that good on kernel info and uses a database that does not
match the real score of the distributions. In some cases, grype reports
a 9.1 score while the distribution gave it a 5.4 and didnt consider
fixing it because its not that high priority or almost impossible to
exploit.

so we are stuck with grype stopping the release while there are no plans
on fixing those.

There is some other times that its totally out of our control and
probably not even affecting them, like the k3s http golang vuln which is
probably not possible on k3s but its being reported as critical.

This is only a workaround but we should probably think about dropping
grype if it requires this much handholding and we have to keep adding
entries to the ignore becuase its just not good enough.

Signed-off-by: Itxaka <itxaka@kairos.io>