Bump agent for kcrypt fix for local TPM
This bumps the agent to the same existing version in 3.1.3 but with a
patched kcrypt dep that brings a small fix for encrypting partitions
with local TPM, in which we need to pass the password to the underlying
command.

No other changes are brought up.

This patch downloads the fixed agent version on build and writes it on
top of the existing one

Signed-off-by: Itxaka <itxaka@kairos.io>